requirements.txt
asgiref==3.8.1
Django==4.2.17
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
importlib_metadata==8.5.0
Markdown==3.7
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
zipp==3.21.0
